refactor db_mongodb.py: refactor update_role_to_db

+++ db/db_mongodb.py
@@ -86,15 +86,20 @@ async def update_role_to_db(*args, **kwargs):
     user_data = kwargs.get('users_data')
     update = kwargs.get('update')
     message = kwargs.get('message')
+    new_role = kwargs.get('role')
 
     collection_name_group_user_role = 'group_user_role'
     collection_group_user_role = db[collection_name_group_user_role]
 
     if user_id:
-        print('update_role_to_db FROM USER_ID')
-        member = await bot.get_chat_member(chat_id, user_id)
-        role = member.status
-        count = await collection_group_user_role.count_documents({'user_id': user_id, 'chat_id': chat_id})
+        if new_role:
+            member = await bot.get_chat_member(chat_id, user_id)
+            role = new_role
+            count = await collection_group_user_role.count_documents({'user_id': user_id, 'chat_id': chat_id})
+        else:
+            member = await bot.get_chat_member(chat_id, user_id)
+            role = member.status
+            count = await collection_group_user_role.count_documents({'user_id': user_id, 'chat_id': chat_id})
 
         user_role = {'user_id': user_id,
                      'username': member.user.username,  # TODO: delete
@@ -164,25 +169,27 @@ async def update_role_to_db(*args, **kwargs):
 
             await add_data_to_db(collection_name_group_user_role, group_user_role)
 
+    elif user_data:
+        count = await collection_group_user_role.count_documents({'user_id': user_data.user.id, 'chat_id': chat_id})
+        user_role = {'user_id': user_data.user.id,
+                     'username': user_data.user.username,  # TODO: delete
+                     'chat_id': chat_id,
+                     'chat_name': 'chat_name',  # TODO: delete
+                     'role': user_data.status
+                     }
+        if count == 0:
+            await add_data_to_db(collection_name_group_user_role, user_role)
+        elif count == 1:
+            filter_update = {'user_id': user_data.user.id, 'chat_id': chat_id}
+            update_role = {'$set': {'role': user_data.status}}
+            collection_group_user_role.update_one(filter_update, update_role)
+        else:
+            delete_filter = {'user_id': user_data.user.id, 'chat_id': chat_id}
+            collection_group_user_role.delete_many(delete_filter)
+            await add_data_to_db(collection_name_group_user_role, user_role)
+
     else:
-        for line in user_data:
-            count = await collection_group_user_role.count_documents({'user_id': line.user.id, 'chat_id': chat_id})
-            user_role = {'user_id': line.user.id,
-                         'username': line.user.username,
-                         'chat_id': chat_id,
-                         'chat_name': 'chat_name',
-                         'role': line.status
-                         }
-            if count == 0:
-                await add_data_to_db(collection_name_group_user_role, user_role)
-            elif count == 1:
-                filter_update = {'user_id': line.user.id, 'chat_id': chat_id}
-                update_role = {'$set': {'role': line.status}}
-                collection_group_user_role.update_one(filter_update, update_role)
-            else:
-                delete_filter = {'user_id': line.user.id, 'chat_id': chat_id}
-                collection_group_user_role.delete_many(delete_filter)
-                await add_data_to_db(collection_name_group_user_role, user_role)
+        print('ты что-то не предусмотрел')
 
 
 async def save_user_to_db_users(chat_id: int, users_data: list):
